+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+x.x.x (x-x-x)
+-------------------
+
+* Updated commands for deployments and environment variables to have no limit by default.
+* Changed --all option to --all-envs for deployments and environment variables to avoid confusion.
+* Display warnings related to the --limit option at the bottom of the rendered content for deployments and environment variables.
+
 3.17.1 (2023-03-09)
 -------------------